+++ doc/rst/MochiKit/VersionHistory.rst
@@ -1,5 +1,7 @@
 2006-XX-XX      v1.4
 
+- MochiKit.Base.evalJSON will now evaluate JSON that is wrapped in a
+  ``/* comment block */``.
 - MochiKit.Color's HSV to RGB conversion code fixed to return the correct
   value (hsv.v for RGB values) in cases where saturation === 0.
 - doXHR workaround for a Mozilla bug when calling XMLHttpRequest in certain